change column type dropdown name
