+++ frontend/src/metadata/components/popover/column-popover/column-type-dropdown-menu/index.css
